Add condition for mr

+++ docs/source/conf.py
@@ -197,7 +197,8 @@ def setup(app):
     path_src = os.path.abspath(".")
     path = os.path.abspath("../scripts")
     print(path, path_src)
-    shutil.rmtree(os.path.join(path_src, "source/scripts"))
+    if os.path.isdir(os.path.join(path_src, "source/scripts")):
+        shutil.rmtree(os.path.join(path_src, "source/scripts"))
     os.mkdir(os.path.join(path_src, "source/scripts"))
     with open("source/scripts/modules.rst", "w") as m:
         m.write("Scripts\n")
